Defeat DBBC3 mcast time-outs if tpicd is stopped
Typically, this is during setup, which will make the DBBC3 busy
and may delay multi-cast. The idea is to give fewer spurious
errors, but still give feedback at other times on whether the
multicast is working. A nice bonus was that it got rid of the
forever loop in read_mcast().